Json-based
Alterations to data_breaches.csv to more useable date format.

R code for constructing a JSON for future use in nesting. Retains state, type of breach, number of records affected, and the date. Remove items with 0 records exposed and no associated state.

Added a test region in main.js for creating a reactive scatterplot to display data for the selected state. Temporary position on page.