Hopefully the last one
Both visualizations are now on the same line, when the window is wide enough to support that. Numerous CSS changes to accomodate that, superfluous heading on map commented out.

Chart legend now horizontal beneath the chart.

Dropdown to select the type of breach displayed. Persists over selecting a new state. This is only affecting visibility, not the actual presence. All points except the displayed are transparent and can still be hovered over to display tooltip. With enough time, a better solution could be implemented to remove those points entirely.

Updated toJson.R to mark values of 0 as 1.001, making them indistinguishable from values of 1, but now the tooltip hover will indicate whether it is a genuien value of 1 or an unknown.

Corrected tooltip so that "1 Record: is displayed rather than "1 Records".

State specific information is now in the title of the plot rather than a separate HTML element.

Updated README.md to include the plot as a feature.